+++ examples/print.rs
@@ -1,3 +1,9 @@
+use std::io::stdin;
+
+use termal::reset_terminal;
+
 fn main() {
-    println!("\x1b[53moverline\x1b[0m");
+    println!("\x1b[52mhello");
+    _ = stdin().read_line(&mut String::new());
+    reset_terminal();
 }