+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Add `ValueOrMut` to make `IoProvider` possible.
 - Add `SS3` code.
 - Add code `OVERLINE`, also add to print macros as `overline`/`ol`.
+- Add codes to enable and disable inverse color in whole terminal.
 + Move around some logic around sixels.
 + Refactor readers.
 + Split into features.